llama: Fix the KV cache quants q4_0 and q8_0 lead server abort in large context chat.
This fix include two parts which cause the problem, first the
context switch need a correct operation for the q4_0 to f32 temp
cache tensor. So I added new function in ggmal_cuda_cpy and
ggml_cuda_dup to process with it.
The second part happened in ggml_compute_forward_dup operation
which also need a correct way to process with q4_0 to f32 and q8_0
to f32, I add them too.